# src/backend/WebApi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Example API",
    "description": "An ASP.NET Core Web API example instance",
    "version": "v1"
  },
  "paths": {
    "/api/v1/Blogging/blog": {
      "get": {
        "tags": [
          "Blogging"
        ],
        "summary": "Gets a list of all blogs.",
        "operationId": "GetBlogs",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Blogging"
        ],
        "summary": "Creates a new blog or updates an existing one.",
        "operationId": "PostBlog",
        "requestBody": {
          "description": "The blog data transfer object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BlogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/Blogging/blog/{id}": {
      "get": {
        "tags": [
          "Blogging"
        ],
        "summary": "Gets a specific blog by ID.",
        "operationId": "GetBlog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the blog.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/Blogging/blog/{blogId}/posts": {
      "get": {
        "tags": [
          "Blogging"
        ],
        "summary": "Gets a list of posts related to a specific blog.",
        "operationId": "GetPosts",
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "description": "The ID of the blog",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/Blogging/post/{id}": {
      "get": {
        "tags": [
          "Blogging"
        ],
        "summary": "Gets a specific post by ID.",
        "operationId": "GetPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the post",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/Blogging/post": {
      "post": {
        "tags": [
          "Blogging"
        ],
        "summary": "Creates a new post.",
        "operationId": "PostPost",
        "requestBody": {
          "description": "The post data transfer object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/Chat/stream": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "SSE stream — keeps the connection open and pushes chat messages as `data: {json}\\n\\n` events.",
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/Chat/messages": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Publishes a chat message. The author is taken from the JWT claims.",
        "requestBody": {
          "description": "Message body.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageRequest"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/Redis/counter": {
      "get": {
        "tags": [
          "Redis"
        ],
        "summary": "Returns the current hit counter value.",
        "operationId": "GetCounter",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedisCounter"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Redis"
        ],
        "summary": "Increments the hit counter and returns the new value.",
        "operationId": "IncrementCounter",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedisCounter"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Redis"
        ],
        "summary": "Resets the hit counter to zero.",
        "operationId": "ResetCounter",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedisCounter"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/SendMessage": {
      "get": {
        "tags": [
          "SendMessage"
        ],
        "summary": "Sends a message using the MessageSender service.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringGenericValue"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/Service/ping": {
      "get": {
        "tags": [
          "Service"
        ],
        "summary": "Returns ok",
        "operationId": "Ping",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringGenericValue"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/Service/version": {
      "get": {
        "tags": [
          "Service"
        ],
        "summary": "Returns version",
        "operationId": "Version",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionInformation"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/Service/status": {
      "get": {
        "tags": [
          "Service"
        ],
        "summary": "Returns the connection status of downstream services (PostgreSQL and RabbitMQ).",
        "operationId": "Status",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceStatus"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BlogDto": {
        "required": [
          "title",
          "url"
        ],
        "type": "object",
        "properties": {
          "blogId": {
            "type": "integer",
            "description": "Gets or sets the unique identifier for the blog.",
            "format": "int32",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "Gets or sets the title of the blog.",
            "nullable": true,
            "example": "My blog title"
          },
          "url": {
            "type": "string",
            "description": "Gets or sets the URL of the blog.",
            "nullable": true,
            "example": "https://localhost/a-test-blog"
          }
        },
        "additionalProperties": false,
        "description": "Blog DTO"
      },
      "ChatMessageRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Message body.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request body for sending a chat message."
      },
      "ConnectionStatus": {
        "required": [
          "connected"
        ],
        "type": "object",
        "properties": {
          "connected": {
            "type": "boolean",
            "description": "Whether the connection is available.",
            "example": true
          },
          "error": {
            "type": "string",
            "description": "Error message if the connection failed, otherwise null.",
            "nullable": true,
            "example": "Connection refused"
          }
        },
        "additionalProperties": false,
        "description": "Represents the status of a single service connection."
      },
      "PostDto": {
        "required": [
          "blogId",
          "content",
          "title"
        ],
        "type": "object",
        "properties": {
          "postId": {
            "type": "integer",
            "description": "Gets or sets the unique identifier for the post.",
            "format": "int32",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "Gets or sets the title of the post.",
            "nullable": true,
            "example": "Bob Loblaw's Law Blog"
          },
          "content": {
            "type": "string",
            "description": "Gets or sets the content of the post.",
            "nullable": true,
            "example": "Example blog post content"
          },
          "blogId": {
            "type": "integer",
            "description": "Gets or sets the unique identifier of the blog to which the post belongs.",
            "format": "int32",
            "example": 1
          }
        },
        "additionalProperties": false,
        "description": "Post DTO"
      },
      "RedisCounter": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "description": "The current counter value.",
            "format": "int64",
            "example": 42
          }
        },
        "additionalProperties": false,
        "description": "A simple counter value stored in Redis."
      },
      "ServiceStatus": {
        "required": [
          "postgres",
          "rabbitMq",
          "redis"
        ],
        "type": "object",
        "properties": {
          "postgres": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "rabbitMq": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "redis": {
            "$ref": "#/components/schemas/ConnectionStatus"
          }
        },
        "additionalProperties": false,
        "description": "Represents the status of all service connections."
      },
      "StringGenericValue": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or sets the value stored in the container.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Represents a generic value container for value types."
      },
      "VersionInformation": {
        "type": "object",
        "properties": {
          "constants": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Constants defined in the assembly, if any.",
            "nullable": true,
            "readOnly": true,
            "example": [
              "DEBUG",
              "NET8"
            ]
          },
          "version": {
            "type": "string",
            "description": "The version of the assembly.",
            "nullable": true,
            "readOnly": true,
            "example": "1.0.0"
          },
          "informationalVersion": {
            "type": "string",
            "description": "The informational version of the assembly, which may include additional details.",
            "nullable": true,
            "readOnly": true,
            "example": "1.0.0-dev"
          },
          "environmentName": {
            "type": "string",
            "description": "The name of the environment where the application is running.",
            "nullable": true,
            "readOnly": true,
            "example": "Development"
          }
        },
        "additionalProperties": false,
        "description": "Represents version information for an assembly, including constants, environment name, version, and informational\nversion."
      }
    },
    "securitySchemes": {
      "oidc": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://auth.wsh.no/api/oidc/authorization",
            "tokenUrl": "https://auth.wsh.no/api/oidc/token",
            "scopes": {
              "openid": "OpenID",
              "profile": "Profile",
              "email": "Email",
              "groups": "Groups",
              "offline_access": "Offline access"
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Blogging"
    },
    {
      "name": "Chat"
    },
    {
      "name": "Redis"
    },
    {
      "name": "SendMessage"
    },
    {
      "name": "Service"
    }
  ]
}